+++ CHANGELOG.md
@@ -24,6 +24,9 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - gen_class.Dnf.\_\_init\_\_: Replaced platform module with distro module.
 - Documentation changes.
 
+### Removed
+- gen_class.KeyCaseInsensitiveDict.has_key: Dictionarys has_key has been removed in Python 3.
+
 
 ## [3.0.0] - 2023-11-07 (Pulled due to unable to work on CentOS 7 and Python 2.7))
 Breaking Changes